Inline the tool schema generator into func_metadata
func_metadata already defined StrictJsonSchema; keep the SEP-2106 external-$ref
rejection there instead of a separate _schema_generator module.
