+++ tests/server/mcpserver/test_tool_manager.py
@@ -10,8 +10,7 @@ from mcp.server.context import LifespanContextT, RequestT
 from mcp.server.mcpserver import Context, MCPServer
 from mcp.server.mcpserver.exceptions import ToolError
 from mcp.server.mcpserver.tools import Tool, ToolManager
-from mcp.server.mcpserver.utilities._schema_generator import ExternalSchemaRefError
-from mcp.server.mcpserver.utilities.func_metadata import ArgModelBase, FuncMetadata
+from mcp.server.mcpserver.utilities.func_metadata import ArgModelBase, ExternalSchemaRefError, FuncMetadata
 from mcp.types import CallToolResult, TextContent, ToolAnnotations
